GoogleChirp/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
pydub==0.25.1
python-multipart==0.0.6
sox==1.4.1
google-api-core==2.11.1
google-auth==2.23.4
google-cloud-core==2.3.3
protobuf==4.25.1 